pyproject.toml
[tool.poetry]
name = "django-ninja-619-reproduction"
version = "0.1.0"
description = ""
authors = ["kazuki.morozumi <kazuki.m777@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2 = "^2.9.5"
django-ninja = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
